sophia-vclock: add vclock.signature; signature txn support

+++ src/box/recovery.h
@@ -165,7 +165,7 @@ void recover_snap(struct recovery_state *r);
 void recovery_follow_local(struct recovery_state *r, ev_tstamp wal_dir_rescan_delay);
 void recovery_finalize(struct recovery_state *r, int rows_per_wal);
 
-int wal_write(struct recovery_state *r, struct xrow_header *packet);
+int64_t wal_write(struct recovery_state *r, struct xrow_header *packet);
 
 void recovery_setup_panic(struct recovery_state *r, bool on_snap_error, bool on_wal_error);
 void recovery_apply_row(struct recovery_state *r, struct xrow_header *packet);
